Fix reconnect not working issue
Fixing the reconnect, since it was not working/reconnecting to TCP Client.
Running the main Reconnect() methode in inside a Task to apply an Task.Delay of about 5 ms, wait for it to complete and then reconnect(). This should give the programm a little buffer so that the necessary rest can be executed.

For TCP Client ListenerTask, also check if the input is null, if so Send a PING to the Server to get new Client?.Connected information, since it only shows the last state and it refreshes with a PING. (IsConnected = true, while the private _isConnected from TCP Client = false, and _isDisconnected = true)