Fixing a rare reconnecting issue, increase Reconnect() delay
This should fix a rare reconnecting issue, where the IClient IsConnected will never be true and will not try to reconnect since lastState and IsConnected is allways true which continues the while loop. Even if it would continue, it wont set needsReconnect to true because _stopServices stays true.

To fix it i added a counter, when the counter hits 50 (which means the IClient isnt connected since 10 Seconds, it will do a Reconnect(), else it would reset the counter

And i increased the delay on the Reconnect() method from 5ms to 20ms to give it a little more time.